HS-42-02: global settings completion — retire the interim drawer
The "History → Settings" move is complete. The interim "Global settings are
consolidating here… lands in HS-30-08… Open History → Settings" drawer is gone;
the gear now opens a real, shell-level /settings page.

- web/src/pages/settings.astro + settings-app.js: a Signal-styled global
  settings surface (Appearance & UI / Core / Cloud intel cards, a Local chip,
  inline validation, a sticky save bar) backed by GET/PUT /api/settings via a
  dedicated settingsApp() factory (load + validator ported verbatim).
- AppLayout.astro: the whole interim settings drawer + its copy + script + CSS
  deleted; the legacy #settings deep link redirects to /settings.
- TopNav.astro: the gear is now an <a href="/settings"> (was the drawer button).
- holdspeak/web/routes/pages.py: /settings serves the built settings page (was
  aliased to history), with a graceful not-built fallback.
- history.astro / history-app.js: the Settings tab + its ~220-line template are
  removed; a read-only loadSettings() stays (the intel-queue alert getters read
  settings.meeting.*); the tab-only saveSettings/validate/state are removed.

Verified live (Playwright against the real server): set History lines = 37 in
the UI -> Save -> reload -> the field AND config.json on disk both show 37.
Screenshot: evidence/settings_page.png.

Tests: +test_web_settings_page.py (route serves the page build-agnostically; the
interim-drawer markers are gone from live source; the gear links to /settings).
3 settings-as-history-tab tests updated in lockstep. Full suite 2286 passed, 16
skipped. Bundle rebuilt; 0 _built/ tracked.

Story HS-42-02 -> done; evidence-story-02.md; phase status + roadmap README.
