+++ pm/roadmap/holdspeak/phase-42-first-run-delight/current-phase-status.md
@@ -1,13 +1,18 @@
 # Phase 42 — First-Run Delight & Daily Confidence
 
-**Status:** IN PROGRESS (1/8 stories). Opened 2026-06-06. Direction chosen by the
+**Status:** IN PROGRESS (2/8 stories). Opened 2026-06-06. Direction chosen by the
 user: the product depth is now ahead of the product **entrance** — make
 **arrival** stellar. A user should go from fresh clone to a verified first
 dictation, with visible privacy/trust state and **zero file editing**, inside one
 guided local cockpit.
 
-**Last updated:** 2026-06-06 (**HS-42-01 done** — the setup-state contract +
-`first_run` milestone. `holdspeak/setup_status.py::build_setup_status` composes
+**Last updated:** 2026-06-06 (**HS-42-02 done** — the interim
+"consolidating / History → Settings" drawer retired into a real shell-level
+**`/settings`** page (Signal cards + sticky save bar); gear→`/settings`; History
+settings tab removed; server route serves the built page; save round-trip proven
+live; +3 tests, 3 settings-as-tab tests updated in lockstep. **HS-42-01 done** —
+the setup-state contract + `first_run` milestone.
+`holdspeak/setup_status.py::build_setup_status` composes
 one `GET /api/setup/status` snapshot as an **adapter** over the existing sources
 (`collect_doctor_checks()` 1:1 → sections; `intel_egress_posture()` + config →
 trust; `detect_presence_platform()` → presence) with a single `primary_action`
@@ -116,7 +121,7 @@ launch and a healthy returning user skips straight to work.
 | ID | Story | Status | Story file | Evidence |
 |---|---|---|---|---|
 | HS-42-01 | Setup-state contract + `first_run` milestone | done | [story-01-setup-state-contract.md](./story-01-setup-state-contract.md) | [evidence-story-01.md](./evidence-story-01.md) |
-| HS-42-02 | Global settings completion | backlog | [story-02-global-settings-completion.md](./story-02-global-settings-completion.md) | — |
+| HS-42-02 | Global settings completion | done | [story-02-global-settings-completion.md](./story-02-global-settings-completion.md) | [evidence-story-02.md](./evidence-story-02.md) |
 | HS-42-03 | Welcome / Setup route + CLI nudge | backlog | [story-03-welcome-setup-route.md](./story-03-welcome-setup-route.md) | — |
 | HS-42-04 | Guided first dictation test (real app) | backlog | [story-04-guided-first-dictation.md](./story-04-guided-first-dictation.md) | — |
 | HS-42-05 | Trust & Privacy — ambient chip + panel | backlog | [story-05-trust-privacy-panel.md](./story-05-trust-privacy-panel.md) | — |
@@ -136,9 +141,14 @@ sources). Resequenced from the proposal so the shell debt (HS-42-02) is retired
 backend spine: `build_setup_status` (an adapter over the doctor/readiness/egress/
 presence sources, cheap via `skip_network`), the durable `first_run` milestone
 (`db.milestones`, survives restart), and `GET /api/setup/status`. +22 tests;
-suite **2283/16**. Next: **HS-42-02** (global settings completion) — retire the
-`AppLayout` "consolidating / History → Settings" interim drawer into a real
-`/settings` surface so the welcome route + trust chip land on a finished shell.
+suite **2283/16**. **HS-42-02 shipped (2026-06-06)** — the interim
+"consolidating / History → Settings" drawer is **retired** into a real,
+shell-level **`/settings`** page (Signal cards: Appearance/Core/Cloud-intel,
+local chip, sticky save bar), the gear links to it, the History settings tab is
+removed (read-only `loadSettings` kept for the intel-alert getters), and the
+server `/settings` route serves the built page. **Save round-trip proven live**
+(set→save→reload→disk all 37) + a screenshot; +3 tests, 3 old settings-as-tab
+tests updated in lockstep. Next: **HS-42-03** (welcome/setup route + CLI nudge).
 
 ## Active risks
