+++ tests/integration/test_web_built_mount.py
@@ -136,14 +136,15 @@ def test_identity_layer_assets_serve(test_client: TestClient) -> None:
 
 
 def test_legacy_routes_still_serve(test_client: TestClient) -> None:
-    # All five route surfaces now serve from the Astro build:
-    # / (HS-10-06), /activity (HS-10-07), /history + /settings
-    # (HS-10-08), /dictation + /docs/dictation-runtime (HS-10-09).
+    # All route surfaces serve from the Astro build:
+    # / (HS-10-06), /activity (HS-10-07), /history (HS-10-08),
+    # /settings (its own page since HS-42-02), /dictation +
+    # /docs/dictation-runtime (HS-10-09).
     for path, marker in [
         ("/", "HoldSpeak"),
         ("/activity", "Local activity"),
         ("/history", "HoldSpeak History"),
-        ("/settings", "HoldSpeak History"),
+        ("/settings", "Every global HoldSpeak setting"),
         ("/dictation", "HoldSpeak Dictation"),
         ("/docs/dictation-runtime", "Dictation runtime setup"),
     ]:
